- Packed some display variables a bit - Y in the projection seems to be reversed. Making it positive Up, reversing the modelview translation vector Y component. I need to see why this needs to be done...

+++ engine/gfx/gfx.h
@@ -27,7 +27,7 @@ typedef enum
 	DF_LIGHTS = 1 << 2,
 	DF_ALL = ~0
 }DIRTYFLAGS;
-extern uint32 g_dirtyFlags;
+extern uint16 g_dirtyFlags;
 #define DF_CHK(x) ((g_dirtyFlags & (x)) != 0)
 #define DF_SET(x) g_dirtyFlags |= (x)
 #define DF_INV(x) g_dirtyFlags &= ~(x)
